# server/protocols/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Camera API",
    "description": "API specification for communication between edge cameras and server",
    "version": "1.0.0",
    "contact": {
      "name": "AI Camera Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.aicamera.com/api/v1",
      "description": "Production server"
    }
  ],
  "paths": {
    "/detection": {
      "post": {
        "summary": "Submit detection data",
        "description": "Submit license plate detection data from edge camera",
        "tags": ["Detection"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionData"
              },
              "example": {
                "camera_id": "cam-001",
                "timestamp": "2024-01-15T10:30:00Z",
                "license_plate": "ABC123",
                "confidence": 0.95,
                "image_url": "https://storage.example.com/images/detection_001.jpg",
                "location": {
                  "latitude": 40.7128,
                  "longitude": -74.0060
                },
                "vehicle_info": {
                  "make": "Toyota",
                  "model": "Camry",
                  "color": "Silver",
                  "type": "Sedan"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Detection data received successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get detection history",
        "description": "Retrieve detection history with optional filters",
        "tags": ["Detection"],
        "parameters": [
          {
            "name": "camera_id",
            "in": "query",
            "description": "Filter by camera ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "license_plate",
            "in": "query",
            "description": "Filter by license plate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Start date for filtering (ISO 8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "End date for filtering (ISO 8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to return",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detection history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DetectionData": {
        "type": "object",
        "required": ["camera_id", "timestamp", "license_plate", "confidence"],
        "properties": {
          "camera_id": {
            "type": "string",
            "description": "Unique identifier for the camera"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of detection"
          },
          "license_plate": {
            "type": "string",
            "pattern": "^[A-Z0-9]{1,10}$",
            "description": "Detected license plate number"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Detection confidence score (0-1)"
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to the captured image"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "vehicle_info": {
            "$ref": "#/components/schemas/VehicleInfo"
          }
        }
      },
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "description": "Latitude coordinate"
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "description": "Longitude coordinate"
          }
        }
      },
      "VehicleInfo": {
        "type": "object",
        "properties": {
          "make": {
            "type": "string",
            "description": "Vehicle manufacturer"
          },
          "model": {
            "type": "string",
            "description": "Vehicle model"
          },
          "color": {
            "type": "string",
            "description": "Vehicle color"
          },
          "type": {
            "type": "string",
            "description": "Vehicle type (e.g., Sedan, SUV, Truck)"
          }
        }
      },
      "DetectionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for the detection record"
          }
        }
      },
      "DetectionListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetectionData"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}